refactoring handlers into sub-components
