+++ .vscode/settings.json
@@ -1,6 +1,6 @@
 {
     "[python]": {
-      "editor.defaultFormatter": "ms-python.black-formatter",
+      "editor.defaultFormatter": "mikoz.black-py",
       "editor.formatOnSave": true
     }
   }
