+++ src/cli/handlers/jazz.py
@@ -1,0 +1,20 @@
+import os
+from datetime import datetime
+
+from colorama import Fore
+
+from genres.jazz.tasks import generate
+
+
+def handle_lead_sheet(args):
+    print(Fore.CYAN + f"\n🎼 Creating MusicXML from {args.input}...")
+    os.makedirs("generated/musicxml", exist_ok=True)
+    timestamp = datetime.now().strftime("%Y-%m-%d_%H%M%S")
+    args.output = f"generated/musicxml/lead_sheet_{timestamp}.musicxml"
+    generate.generate_jazz_lead_sheet(args.input, args.output)
+    print(Fore.YELLOW + f"\n💾 MusicXML saved to {args.output}")
+
+
+handler_map = {
+    "lead-sheet": handle_lead_sheet,
+}
